requirements.txt
fs
azure-storage-blob
requests
